requirements.txt
# local package
-e .

# external requirements. 
#TODO Should be updated to reflect versioning
click
docutils==0.15
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
xlrd
tldextract
pandas
numpy
pandas-profiling
nltk
pygam
matplotlib
graphviz
pydotplus
flask
pickle5
scikit-learn
mlflow
pytest
requests
kafka-python
logger